MS-1 v8 компилируется без предупреждений

+++ sources/MS-1/src/Modules/BME280/BME280.cpp
@@ -82,7 +82,7 @@ bool BME280::GetMeasures(float* temp, float* pressure, float* humidity)
         return false;
     }
 
-    timeNext += TIME_MEASURE + (std::rand() % 100);
+    timeNext += TIME_MEASURE + (uint)(std::rand() % 100);
 
 #ifdef IN_MODE_TEST
